pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamchunk"
version = "2.0.0"
description = "Adaptive data stream chunker with CPU-aware parallel processing for real-time ETL pipelines"
requires-python = ">=3.8"
dependencies = [
    "psutil>=5.9",
    "numpy>=1.21",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
pandas   = ["pandas>=1.3"]
kafka    = ["kafka-python>=2.0"]
database = ["sqlalchemy>=1.4"]
metrics  = ["prometheus-client>=0.14"]
async    = ["aiofiles>=0.8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
